% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commensurate.bin.R
\name{commensurate.bin}
\alias{commensurate.bin}
\title{Simulation study of hybrid control design with commensurate power prior:
binary outcome}
\usage{
commensurate.bin(
  n.CT, n.CC, n.EC,
  out.prob.CT, out.prob.CC, driftOR,
  cov.C, cov.cor.C, cov.effect.C,
  cov.E, cov.cor.E, cov.effect.E,
  method, chains=2, iter=4000, warmup=floor(iter/2), thin=1,
  alternative="greater", sig.level=0.025, nsim)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment.}

\item{n.CC}{Number of patients in concurrent control.}

\item{n.EC}{Number of patients in external control.}

\item{out.prob.CT}{True probability of outcome in concurrent treatment.}

\item{out.prob.CC}{True probability of outcome in concurrent control.}

\item{driftOR}{OR between external control and concurrent control for which
the bias should be plotted.}

\item{cov.C}{List of covariates in concurrent data. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.cor.C}{Matrix of correlation coefficients for covariates in
concurrent data, specified as Gaussian copula parameter.}

\item{cov.effect.C}{Vector of covariate effects in concurrent data,
specified as odds ratio.}

\item{cov.E}{List of covariates in external data. Distribution and
and its parameters need to be specified for each covariate.}

\item{cov.cor.E}{Matrix of correlation coefficients for covariates in
external data, specified as Gaussian copula parameter.}

\item{cov.effect.E}{Vector of covariate effects in external data,
specified as odds ratio.}

\item{method}{List of information borrowing method. \code{"noborrow"} uses
the concurrent data only. \code{"fullborrow"} uses the external control data
without discounting. \code{"cauchy"} uses the commensurate prior to
dynamically borrow the external control data, and the commensurate parameter
is assumed to follow half cauchy distribution. \code{"normal"} uses the
commensurate prior to dynamically borrow the external control data, and the
commensurate parameter is assumed to follow half normal distribution.
\code{"cauchy"} and \code{"normal"} require to specify the scale parameter of
half cauchy and half normal distribution respectively.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=2}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=4000}.}

\item{warmup}{Number of warmup (aka burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.025}.}

\item{nsim}{Number of simulated trials. The default value is \code{nsim=10}.}
}
\value{
\item{reject}{\code{TRUE} when significant; otherwise \code{FALSE}.}
\item{theta}{Posterior mean, median, and sd of log odds ratio.}
}
\description{
Bayesian dynamic information borrowing with commensurate power prior is
implemented for hybrid control design, where the concurrent control is
augmented by historical control. No borrowing and full borrowing are also
implemented. The binary outcome is applicable.
}
\examples{
n.CT  <- 50
n.CC  <- 50
n.EC  <- 50

out.prob.CT <- 0.2
out.prob.CC <- 0.2
driftOR     <- 1.0

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.E <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.E <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.E <- c(0.1,0.1)

method <- list(list(prior="noborrow"),
               list(prior="normal",scale=0.5))

nsim <- 5

commensurate.bin(
  n.CT=n.CT, n.CC=n.CC, n.EC=n.EC,
  out.prob.CT=out.prob.CT, out.prob.CC=out.prob.CC, driftOR=driftOR,
  cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
  cov.E=cov.E, cov.cor.E=cov.cor.E, cov.effect.E=cov.effect.E,
  method=method, nsim=nsim)
}
