% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commensurate.summary.R
\name{commensurate.summary}
\alias{commensurate.summary}
\title{Summarizing simulation study results}
\usage{
commensurate.summary(simres,method.lab,t.theta)
}
\arguments{
\item{simres}{List of simulation results, including reject and theta.}

\item{method.lab}{List of information borrowing method.}

\item{t.theta}{True value of treatment effect.}
}
\description{
Simulation study results are summarized.
}
