% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen}
\alias{datagen}
\title{Generate multivariate correlated data}
\usage{
datagen(margdist, corvec, nsim)
}
\arguments{
\item{margdist}{List of distributions to be used for the data generation.}

\item{corvec}{Vector of Gaussian copula correlation parameter.}

\item{nsim}{Number of simulation.}
}
\value{
Data having the specified distribution.
}
\description{
Multivariate correlated data are generated. Gaussian copula is used to
specify the correlation between variables. Any probability distributions
available in R STAT is applicable.
}
