% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBayesborrow-package.R
\name{psBayesborrow-package}
\alias{psBayesborrow-package}
\title{The 'psBayesborrow' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
