% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psborrow.bin.R
\name{psborrow.bin}
\alias{psborrow.bin}
\title{Simulation study of hybrid control design with Bayesian dynamic borrowing
incorporating propensity score matched external control: binary outcome}
\usage{
psborrow.bin(
  n.CT, n.CC, n.ECp, n.EC,
  out.prob.CT, out.prob.CC, driftOR,
  cov.C, cov.cor.C, cov.effect.C,
  cov.EC, cov.cor.EC, cov.effect.EC,
  method.psest="glm", method.pslink="logit",
  method.whomatch, method.matching, method.psorder,
  method.borrow,
  chains=2, iter=4000, warmup=floor(iter/2), thin=1,
  alternative="greater", sig.level=0.025, nsim, seed=NULL)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment.}

\item{n.CC}{Number of patients in concurrent control.}

\item{n.ECp}{Number of patients in external control pool.}

\item{n.EC}{Number of patients in external control.}

\item{out.prob.CT}{True probability of outcome in concurrent treatment.}

\item{out.prob.CC}{True probability of outcome in concurrent control.}

\item{driftOR}{OR between external control and concurrent control for which
the bias should be plotted.}

\item{cov.C}{List of covariate distributions for concurrent treatment and
control.}

\item{cov.cor.C}{Matrix of correlation coefficients for each pair of
covariate for concurrent treatment and control, specified as Gaussian copula
parameter.}

\item{cov.effect.C}{Vector of covariate effects for concurrent treatment and
control, specified as hazard ratio.}

\item{cov.EC}{List of covariate distributions for external control.}

\item{cov.cor.EC}{Matrix of correlation coefficients for each pair of
covariate for external control, specified as Gaussian copula parameter.}

\item{cov.effect.EC}{Vector of covariate effects for external control
control, specified as hazard ratio.}

\item{method.psest}{Method of estimating the propensity score. Allowable
options include, for example, \code{"glm"} for generalized linear model
(e.g., logistic regression); \code{"gam"} for generalized additive model;
\code{"gbm"} for generalized boosted model; \code{"lasso"} for lasso
regression; \code{"rpart"} for classification tree. The default value is
\code{method.psest="glm"}.}

\item{method.pslink}{Link function used in estimating the propensity score.
Allowable options depend on the specific \code{method.psest} value specified.
The default value is \code{method.pslink="logit"}, which, along with
\code{method.psest="glm"}, identifies the default method as logistic
regression.}

\item{method.whomatch}{Options of who to match. Allowable options include
\code{conc.contl} matching concurrent control to external control pool;
\code{conc.treat} matching concurrent treatment to external control pool;
\code{conc.all} matching concurrent treatment plus concurrent control to
external control pool; \code{treat2contl} matching concurrent treatment to
concurrent control plus external control pool.}

\item{method.matching}{Matching method. Allowable options include
\code{"optimal"} for optimal matching; \code{"nearest"} for nearest neighbor
matching without replacement.}

\item{method.psorder}{Order that the matching takes place. Allowable options
include \code{"largest"}, where matching takes place in descending order of
propensity score; \code{"smallest"}, where matching takes place in ascending
order of propensity score; \code{"random"}, where matching takes place in a
random order; \code{"data"}, where matching takes place based on the order
of units in the data.}

\item{method.borrow}{List of information borrowing method. \code{"noborrow"}
uses the concurrent data only. \code{"fullborrow"} uses the external control
data without discounting. \code{"cauchy"} uses the commensurate prior to
dynamically borrow the external control data, and the commensurate parameter
is assumed to follow half-Cauchy distribution. \code{"normal"} uses the
commensurate prior to dynamically borrow the external control data, and the
commensurate parameter is assumed to follow half-normal distribution.
\code{"cauchy"} and \code{"normal"} require to specify the scale parameter of
half-Cauchy and half-normal distribution respectively.}

\item{chains}{Number of Markov chains in MCMC sampling. The default value is
\code{chains=2}.}

\item{iter}{Number of iterations for each chain (including warmup) in MCMC
sampling. The default value is \code{iter=4000}.}

\item{warmup}{Number of warmup (burnin) iterations per chain in MCMC
sampling. The default value is \code{warmup=floor(iter/2)}.}

\item{thin}{Period for saving samples in MCMC sampling. The default value
is \code{thin=1}.}

\item{alternative}{Alternative hypothesis to be tested ("greater" or "less").
The default value is \code{alternative="greater"}.}

\item{sig.level}{Significance level. The default value is
\code{sig.level=0.025}.}

\item{nsim}{Number of simulated trials.}

\item{seed}{Random seed. The default value is \code{seed=NULL}.}
}
\value{
\item{reject}{\code{TRUE} when significant; otherwise \code{FALSE}.}
\item{theta}{Posterior mean, median, and sd of log hazard ratio.}
\item{ov.ps}{Overlapping coefficient of propensity score between concurrent
treatment and concurrent/external control}
\item{ov.cov}{Overlapping coefficient of continuous covariate and difference
of proportion of binary covariate between concurrent treatment and
concurrent/external control}
}
\description{
Bayesian dynamic borrowing is implemented for clinical trials with hybrid
control design, where the concurrent control is augmented by external
control. The external control is selected from external control pool using
propensity score matching. Commensurate power prior is used for Bayesian
dynamic borrowing, and half-normal or half-Cauchy commensurate prior is
available. No borrowing and full borrowing can also be implemented.
The binary outcome is applicable.
}
\examples{
n.CT       <- 100
n.CC       <- 50
n.ECp      <- 1000
n.EC       <- 50

out.prob.CT <- 0.2
out.prob.CC <- 0.2
driftOR     <- 1.0

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.EC <- list(list(dist="norm",mean=0,sd=1),
               list(dist="binom",prob=0.4))

cov.cor.EC <- rbind(c(  1,0.1),
                    c(0.1,  1))

cov.effect.EC <- c(0.1,0.1)

method.whomatch <- "conc.treat"
method.matching <- "optimal"
method.psorder  <- NULL

method.borrow <- list(list(prior="noborrow"),
                      list(prior="normal",scale=0.5))

nsim <- 5

psborrow.bin(
  n.CT=n.CT, n.CC=n.CC, n.ECp=n.ECp, n.EC=n.EC,
  out.prob.CT=out.prob.CT, out.prob.CC=out.prob.CC, driftOR=driftOR,
  cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
  cov.EC=cov.EC, cov.cor.EC=cov.cor.EC, cov.effect.EC=cov.effect.EC,
  method.whomatch=method.whomatch,
  method.matching=method.matching, method.psorder=method.psorder,
  method.borrow=method.borrow, nsim=nsim, seed=100)
}
