% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psborrow.summary.R
\name{psborrow.summary}
\alias{psborrow.summary}
\title{Summarizing simulation study results}
\usage{
psborrow.summary(object)
}
\arguments{
\item{object}{List of simulation results.}
}
\description{
Simulation study results are summarized.
}
