% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.simulation.bin.R
\name{trial.simulation.bin}
\alias{trial.simulation.bin}
\title{Simulating a trial with binary outcome}
\usage{
trial.simulation.bin(
 n.CT, n.CC, n.ECp,
 out.prob.CT, out.prob.CC, driftOR,
 cov.C, cov.cor.C, cov.effect.C,
 cov.EC, cov.cor.EC, cov.effect.EC)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment.}

\item{n.CC}{Number of patients in concurrent control.}

\item{n.ECp}{Number of patients in external control pool.}

\item{out.prob.CT}{True probability of outcome in concurrent treatment.}

\item{out.prob.CC}{True probability of outcome in concurrent control.}

\item{driftOR}{OR between external control and concurrent control for which
the bias should be plotted.}

\item{cov.C}{List of covariate distributions for concurrent treatment and
control.}

\item{cov.cor.C}{Matrix of correlation coefficients for each pair of
covariate for concurrent treatment and control, specified as Gaussian copula
parameter.}

\item{cov.effect.C}{Vector of covariate effects for concurrent treatment and
control, specified as hazard ratio.}

\item{cov.EC}{List of covariate distributions for external control.}

\item{cov.cor.EC}{Matrix of correlation coefficients for each pair of
covariate for external control, specified as Gaussian copula parameter.}

\item{cov.effect.EC}{Vector of covariate effects for external control
control, specified as hazard ratio.}
}
\value{
\item{outdata}{Dataset of a simulated trial.}
\item{ncov}{Number of covariate.}
}
\description{
A trial with binary outcome is simulated. The data returned include
the time to event of interest, an indicator of censoring, and covariate for
patients in concurrent treatment and control, and external control pool.
}
\examples{
n.CT       <- 100
n.CC       <- 50
n.ECp      <- 1000

out.prob.CT <- 0.2
out.prob.CC <- 0.2
driftOR     <- 1.0

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.EC <- list(list(dist="norm",mean=0,sd=1),
               list(dist="binom",prob=0.4))

cov.cor.EC <- rbind(c(  1,0.1),
                    c(0.1,  1))

cov.effect.EC <- c(0.1,0.1)

trial.simulation.bin(
   n.CT=n.CT, n.CC=n.CC, n.ECp=n.ECp,
   out.prob.CT=out.prob.CT, out.prob.CC=out.prob.CC, driftOR=driftOR,
   cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
   cov.EC=cov.EC, cov.cor.EC=cov.cor.EC, cov.effect.EC=cov.effect.EC)
}
