% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.simulation.t2e.R
\name{trial.simulation.t2e}
\alias{trial.simulation.t2e}
\title{Simulating a trial with time-to-event outcome}
\usage{
trial.simulation.t2e(
 n.CT, n.CC, nevent.C, n.ECp, nevent.ECp, accrual,
 out.mevent.CT, out.mevent.CC, driftHR,
 cov.C, cov.cor.C, cov.effect.C,
 cov.EC, cov.cor.EC, cov.effect.EC)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment.}

\item{n.CC}{Number of patients in concurrent control.}

\item{nevent.C}{Number of events in concurrent treatment and control.}

\item{n.ECp}{Number of patients in external control pool.}

\item{nevent.ECp}{Number of events in external control pool.}

\item{accrual}{Accrual rate (number of enrolled patients per month).}

\item{out.mevent.CT}{True median time to event in concurrent treatment.}

\item{out.mevent.CC}{True median time to event in concurrent control.}

\item{driftHR}{Hazard ratio between concurrent and external control for
which the bias should be plotted.}

\item{cov.C}{List of covariate distributions for concurrent treatment and
control.}

\item{cov.cor.C}{Matrix of correlation coefficients for each pair of
covariate for concurrent treatment and control, specified as Gaussian copula
parameter.}

\item{cov.effect.C}{Vector of covariate effects for concurrent treatment and
control, specified as hazard ratio.}

\item{cov.EC}{List of covariate distributions for external control.}

\item{cov.cor.EC}{Matrix of correlation coefficients for each pair of
covariate for external control, specified as Gaussian copula parameter.}

\item{cov.effect.EC}{Vector of covariate effects for external control
control, specified as hazard ratio.}
}
\value{
\item{outdata}{Dataset of a simulated trial.}
\item{ncov}{Number of covariate.}
}
\description{
A trial with time-to-event outcome is simulated. The data returned include
the time to event of interest, an indicator of censoring, and covariate for
patients in concurrent treatment and control, and external control pool.
}
\examples{
n.CT       <- 100
n.CC       <- 50
nevent.C   <- 100
n.ECp      <- 1000
nevent.ECp <- 800
accrual    <- 16

out.mevent.CT <- 6
out.mevent.CC <- 6
driftHR       <- 1

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.EC <- list(list(dist="norm",mean=0,sd=1),
               list(dist="binom",prob=0.4))

cov.cor.EC <- rbind(c(  1,0.1),
                    c(0.1,  1))

cov.effect.EC <- c(0.1,0.1)

trial.simulation.t2e(
   n.CT=n.CT, n.CC=n.CC, nevent.C=nevent.C,
   n.ECp=n.ECp, nevent.ECp=nevent.ECp, accrual=accrual,
   out.mevent.CT, out.mevent.CC, driftHR,
   cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
   cov.EC=cov.EC, cov.cor.EC=cov.cor.EC, cov.effect.EC=cov.effect.EC)
}
