% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psBayesborrow-package.R
\docType{package}
\name{psBayesborrow-package}
\alias{psBayesborrow-package}
\alias{psBayesborrow}
\title{The 'psBayesborrow' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
\author{
\strong{Maintainer}: Yusuke Yamaguchi \email{yamagubed@gmail.com}

Authors:
\itemize{
  \item Jun Takeda
  \item Kentaro Takeda
}

}
