% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatch.R
\name{psmatch}
\alias{psmatch}
\title{Propensity score matching}
\usage{
psmatch(
    formula, data, n.EC,
    method.psest="glm", method.pslink="logit",
    method.whomatch, method.matching, method.psorder)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}

\item{n.EC}{Number of patients in external control.}

\item{method.psest}{Method of estimating the propensity score. Allowable
options include, for example, \code{"glm"} for generalized linear model
(e.g., logistic regression); \code{"gam"} for generalized additive model;
\code{"gbm"} for generalized boosted model; \code{"lasso"} for lasso
regression; \code{"rpart"} for classification tree. The default value is
\code{method.psest="glm"}.}

\item{method.pslink}{Link function used in estimating the propensity score.
Allowable options depend on the specific \code{method.psest} value specified.
The default value is \code{method.pslink="logit"}, which, along with
\code{method.psest="glm"}, identifies the default method as logistic
regression.}

\item{method.whomatch}{Options of who to match. Allowable options include
\code{conc.contl} matching concurrent control to external control pool;
\code{conc.treat} matching concurrent treatment to external control pool;
\code{conc.all} matching concurrent treatment plus concurrent control to
external control pool; \code{treat2contl} matching concurrent treatment to
concurrent control plus external control pool.}

\item{method.matching}{Matching method. Allowable options include
\code{"optimal"} for optimal matching; \code{"nearest"} for nearest neighbor
matching without replacement.}

\item{method.psorder}{Order that the matching takes place. Allowable options
include \code{"largest"}, where matching takes place in descending order of
propensity score; \code{"smallest"}, where matching takes place in ascending
order of propensity score; \code{"random"}, where matching takes place in a
random order; \code{"data"}, where matching takes place based on the order
of units in the data.}
}
\value{
\item{subjid.EC}{Subject ID of external control.}
\item{data.cov.ps}{Dataset of a simulated trial with estimated propensity
score.}
}
\description{
Propensity score matching is implemented using methods specified.
}
\examples{
n.CT       <- 100
n.CC       <- 50
nevent.C   <- 100
n.ECp      <- 1000
nevent.ECp <- 800
accrual    <- 16

out.mevent.CT <- 6
out.mevent.CC <- 6
driftHR       <- 1

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.EC <- list(list(dist="norm",mean=0,sd=1),
               list(dist="binom",prob=0.4))

cov.cor.EC <- rbind(c(  1,0.1),
                    c(0.1,  1))

cov.effect.EC <- c(0.1,0.1)

indata <- trial.simulation.t2e(
  n.CT=n.CT, n.CC=n.CC, nevent.C=nevent.C,
  n.ECp=n.ECp, nevent.ECp=nevent.ECp, accrual=accrual,
  out.mevent.CT, out.mevent.CC, driftHR,
  cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
  cov.EC=cov.EC, cov.cor.EC=cov.cor.EC, cov.effect.EC=cov.effect.EC)

n.EC <- 50

method.whomatch <- "conc.treat"
method.matching <- "optimal"
method.psorder  <- NULL

psmatch(
  study~X1+X2, data=indata, n.EC=n.EC,
  method.whomatch=method.whomatch, method.matching=method.matching,
  method.psorder=method.psorder)
}
