% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.simulation.cont.R
\name{trial.simulation.cont}
\alias{trial.simulation.cont}
\title{Simulating a trial with continuous outcome}
\usage{
trial.simulation.cont(
 n.CT, n.CC, n.ECp,
 out.mean.CT, out.sd.CT, out.mean.CC, out.sd.CC, driftdiff, out.sd.EC,
 cov.C, cov.cor.C, cov.effect.C,
 cov.EC, cov.cor.EC, cov.effect.EC)
}
\arguments{
\item{n.CT}{Number of patients in concurrent treatment.}

\item{n.CC}{Number of patients in concurrent control.}

\item{n.ECp}{Number of patients in external control pool.}

\item{out.mean.CT}{True mean of outcome in concurrent treatment.}

\item{out.sd.CT}{True sd of outcome in concurrent treatment.}

\item{out.mean.CC}{True mean of outcome in concurrent control.}

\item{out.sd.CC}{True as of outcome in concurrent control.}

\item{driftdiff}{Difference between external control and concurrent control
for which the bias should be plotted.}

\item{out.sd.EC}{True as of outcome in external control.}

\item{cov.C}{List of covariate distributions for concurrent treatment and
control.}

\item{cov.cor.C}{Matrix of correlation coefficients for each pair of
covariate for concurrent treatment and control, specified as Gaussian copula
parameter.}

\item{cov.effect.C}{Vector of covariate effects for concurrent treatment and
control, specified as hazard ratio.}

\item{cov.EC}{List of covariate distributions for external control.}

\item{cov.cor.EC}{Matrix of correlation coefficients for each pair of
covariate for external control, specified as Gaussian copula parameter.}

\item{cov.effect.EC}{Vector of covariate effects for external control
control, specified as hazard ratio.}
}
\value{
\item{outdata}{Dataset of a simulated trial.}
\item{ncov}{Number of covariate.}
}
\description{
A trial with continuous outcome is simulated. The data returned include
the time to event of interest, an indicator of censoring, and covariate for
patients in concurrent treatment and control, and external control pool.
}
\examples{
n.CT       <- 100
n.CC       <- 50
n.ECp      <- 1000

out.mean.CT <- 0
out.sd.CT   <- 1
out.mean.CC <- 0
out.sd.CC   <- 1
driftdiff   <- 0
out.sd.EC   <- 1

cov.C <- list(list(dist="norm",mean=0,sd=1),
              list(dist="binom",prob=0.4))

cov.cor.C <- rbind(c(  1,0.1),
                   c(0.1,  1))

cov.effect.C <- c(0.1,0.1)

cov.EC <- list(list(dist="norm",mean=0,sd=1),
               list(dist="binom",prob=0.4))

cov.cor.EC <- rbind(c(  1,0.1),
                    c(0.1,  1))

cov.effect.EC <- c(0.1,0.1)

trial.simulation.cont(
  n.CT=n.CT, n.CC=n.CC, n.ECp=n.ECp,
  out.mean.CT=out.mean.CT, out.sd.CT=out.sd.CT,
  out.mean.CC=out.mean.CC, out.sd.CC=out.sd.CC,
  driftdiff=driftdiff, out.sd.EC=out.sd.EC,
  cov.C=cov.C, cov.cor.C=cov.cor.C, cov.effect.C=cov.effect.C,
  cov.EC=cov.EC, cov.cor.EC=cov.cor.EC, cov.effect.EC=cov.effect.EC)
}
